p3_interpolation

Function interpolate_coset

pub fn interpolate_coset<F, EF, Mat>(
    coset_evals: &Mat,
    shift: F,
    point: EF,
) -> Vec<EF>
where
    F: TwoAdicField,
    EF: ExtensionField<F>,
    Mat: Matrix<F>,

Given evaluations of a batch of polynomials over the given coset of the canonical power-of-two subgroup, evaluate the polynomials at point.

This assumes the point is not in the coset, otherwise the behavior is undefined.

The coset_evals must be given in standard (not bit-reversed) order.